Made changes as requested in pull #73
Summary:
- Class renames
- Removed the ability to read out the results of an InlinedOperator/TupleValueSink in a generic format. We can figure out how we want to expose this information at a later time.
