--- src/main/java/com/facebook/presto/operator/inlined/InlinedOperator.java
@@ -1,6 +1,0 @@
-package com.facebook.presto.operator.inlined;
-
-public interface InlinedOperator
-    extends InlinedOperatorWriter, InlinedOperatorReader
-{
-}
