--- src/main/java/com/facebook/presto/operator/inlined/InlinedOperatorReader.java
@@ -1,17 +1,0 @@
-package com.facebook.presto.operator.inlined;
-
-import com.facebook.presto.Range;
-import com.facebook.presto.TupleInfo;
-import com.facebook.presto.block.TupleStream;
-
-public interface InlinedOperatorReader
-{
-    boolean isFinished();
-    TupleInfo getTupleInfo();
-    Range getRange();
-
-    /**
-     * Returns the current result as a TupleStream (can be partial result if not yet finished)
-     */
-    TupleStream getResult();
-}
