--- src/main/java/com/facebook/presto/operator/inlined/InlinedOperatorWriter.java
@@ -1,9 +1,0 @@
-package com.facebook.presto.operator.inlined;
-
-import com.facebook.presto.block.TupleStreamPosition;
-
-public interface InlinedOperatorWriter
-{
-    void process(TupleStreamPosition tupleStreamPosition);
-    void finished();
-}
